+++ lib/puppet/type/shared/extraproperties.rb
@@ -9,6 +9,10 @@ newproperty(:extraproperties, :array_matching => :all) do
     end
   end
 
+  def insync?(is)
+    is.sort == should.sort
+  end
+
 end
 
 def extraproperties