+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # History
 
 ## 1.0.19
-- changed properties of wls_mail_session, no change when there is a different sorting
+- changed property and removed propertyvalues of wls_mail_session, now there is no change detected when there is a different sorting
 - changed users of wls_group, no change when there is a different sorting
+- removed extrapropertiesvalues of wls_datasource, you can use now extraproperties with key=value,key1=value
+- removed extrapropertiesvalues of wls_foreign_server, you can use now extraproperties with key=value,key1=value
 
 ## 1.0.18
 - small fix for the server_template type with the arguments parameter